Added new labels tabel (including migration) and fixed issue with transactions not active.
