+++ app/pom.xml
@@ -138,6 +138,11 @@ SPDX-License-Identifier: Apache-2.0
             <artifactId>openpojo</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-simple</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
@@ -156,6 +161,29 @@ SPDX-License-Identifier: Apache-2.0
                         </goals>
                     </execution>
                     <execution>
+                        <id>build-basex</id>
+                        <goals>
+                            <goal>build</goal>
+                        </goals>
+                        <configuration>
+                            <properties>
+                                <quarkus.profile>prod-basex</quarkus.profile>
+                                <quarkus.package.output-directory>basex-quarkus-app</quarkus.package.output-directory>
+                                <quarkus.docker.dockerfile-jvm-path>src/main/docker/Dockerfile-basex.jvm
+                                </quarkus.docker.dockerfile-jvm-path>
+                                <quarkus.docker.dockerfile-native-path>src/main/docker/Dockerfile-basex.native
+                                </quarkus.docker.dockerfile-native-path>
+                                <quarkus.container-image.tag>${project.version}-basex</quarkus.container-image.tag>
+                                <quarkus.container-image.additional-tags>latest-basex
+                                </quarkus.container-image.additional-tags>
+                            </properties>
+                        </configuration>
+                    </execution>
+                    <!-- Put this build execution as last, because we will use this image to do a basic native check (NativeHealthCheckIT).
+                         This information is written in "target\quarkus-artifact.properties" and used by the integration tests (failsafe).
+                         Sadly there is only 1 file create, so this build overwrites the one from basex.
+                      -->
+                    <execution>
                         <id>build-postgresql</id>
                         <goals>
                             <goal>build</goal>
@@ -171,29 +199,6 @@ SPDX-License-Identifier: Apache-2.0
                                 </quarkus.docker.dockerfile-native-path>
                                 <quarkus.container-image.tag>${project.version}-postgresql</quarkus.container-image.tag>
                                 <quarkus.container-image.additional-tags>latest-postgresql
-                                </quarkus.container-image.additional-tags>
-                            </properties>
-                        </configuration>
-                    </execution>
-                    <!-- Put this build execution as last, because we will use this image to do a basic native check (NativeHealthCheckIT).
-                         This information is written in "target\quarkus-artifact.properties" and used by the integration tests (failsafe).
-                         Sadly there is only 1 file create, so this build overwrites the one from postgresql.
-                      -->
-                    <execution>
-                        <id>build-basex</id>
-                        <goals>
-                            <goal>build</goal>
-                        </goals>
-                        <configuration>
-                            <properties>
-                                <quarkus.profile>prod-basex</quarkus.profile>
-                                <quarkus.package.output-directory>basex-quarkus-app</quarkus.package.output-directory>
-                                <quarkus.docker.dockerfile-jvm-path>src/main/docker/Dockerfile-basex.jvm
-                                </quarkus.docker.dockerfile-jvm-path>
-                                <quarkus.docker.dockerfile-native-path>src/main/docker/Dockerfile-basex.native
-                                </quarkus.docker.dockerfile-native-path>
-                                <quarkus.container-image.tag>${project.version}-basex</quarkus.container-image.tag>
-                                <quarkus.container-image.additional-tags>latest-basex
                                 </quarkus.container-image.additional-tags>
                             </properties>
                         </configuration>
