+++ app/src/main/resources/application-dev-basex.properties
@@ -14,6 +14,10 @@ quarkus.log.category."org.lfenergy.compas.scl.data".level = DEBUG
 quarkus.http.auth.permission.develop-quarkus-services.paths  = /compas-scl-data-service/q/swagger-ui/*,/compas-scl-data-service/index.html
 quarkus.http.auth.permission.develop-quarkus-services.policy = permit
 
+# Exclude the other repository implementations from being scanned
+quarkus.arc.exclude-dependency.postgresql-repository.group-id    = org.lfenergy.compas.scl.data
+quarkus.arc.exclude-dependency.postgresql-repository.artifact-id = repository-postgresql
+
 # Datasource configuration for BaseX (none)
 quarkus.datasource.jdbc                = false
 quarkus.datasource.devservices.enabled = false
