+++ app/src/main/resources/application-prod-basex.properties
@@ -4,9 +4,13 @@
 
 # Production BaseX configuration.
 
+# Exclude the other repository implementations from being scanned
+quarkus.arc.exclude-dependency.postgresql-repository.group-id    = org.lfenergy.compas.scl.data
+quarkus.arc.exclude-dependency.postgresql-repository.artifact-id = repository-postgresql
+
 # Add scanning these dependencies for scanning, also used by native compilation.
-quarkus.index-dependency.scl-data-repository.group-id    = org.lfenergy.compas.scl.data
-quarkus.index-dependency.scl-data-repository.artifact-id = repository-basex
+quarkus.index-dependency.basex-repository.group-id    = org.lfenergy.compas.scl.data
+quarkus.index-dependency.basex-repository.artifact-id = repository-basex
 
 # Datasource configuration for BaseX (none)
 quarkus.datasource.jdbc = false
