+++ app/src/main/resources/application-prod-postgresql.properties
@@ -4,9 +4,13 @@
 
 # Production PostgreSQL configuration.
 
+# Exclude the other repository implementations from being scanned
+quarkus.arc.exclude-dependency.basex-repository.group-id    = org.lfenergy.compas.scl.data
+quarkus.arc.exclude-dependency.basex-repository.artifact-id = repository-basex
+
 # Add scanning these dependencies for scanning, also used by native compilation.
-quarkus.index-dependency.scl-data-repository.group-id    = org.lfenergy.compas.scl.data
-quarkus.index-dependency.scl-data-repository.artifact-id = repository-postgresql
+quarkus.index-dependency.postgresql-repository.group-id    = org.lfenergy.compas.scl.data
+quarkus.index-dependency.postgresql-repository.artifact-id = repository-postgresql
 
 # Datasource configuration for PostgreSQL
 quarkus.datasource.db-kind       = postgresql
