+++ app/src/main/resources/application-test.properties
@@ -4,10 +4,15 @@
 
 # Test configuration.
 
-# Datasource configuration for BaseX (none)
-quarkus.datasource.jdbc                = false
-quarkus.datasource.devservices.enabled = false
+# Exclude the other repository implementations from being scanned
+quarkus.arc.exclude-dependency.basex-repository.group-id    = org.lfenergy.compas.scl.data
+quarkus.arc.exclude-dependency.basex-repository.artifact-id = repository-basex
 
-# Flyway configuration for BaseX (none)
-quarkus.flyway.migrate-at-start = false
+# Datasource configuration for PostgreSQL
+quarkus.datasource.devservices.enabled = true
+quarkus.datasource.db-kind             = postgresql
+
+# Flyway configuration for PostgreSQL
+quarkus.flyway.migrate-at-start = true
+quarkus.flyway.locations        = classpath:org/lfenergy/compas/scl/data/repository/postgresql/db/migration
