--- app/src/main/resources/reflection-config.json.license
@@ -1,3 +1,0 @@
-SPDX-FileCopyrightText: 2021 Alliander N.V.
-
-SPDX-License-Identifier: Apache-2.0
