+++ repository/pom.xml
@@ -45,6 +45,10 @@ SPDX-License-Identifier: Apache-2.0
             <groupId>jakarta.transaction</groupId>
             <artifactId>jakarta.transaction-api</artifactId>
         </dependency>
+        <dependency>
+            <groupId>jakarta.enterprise</groupId>
+            <artifactId>jakarta.enterprise.cdi-api</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>org.slf4j</groupId>
