+++ repository/src/main/java/org/lfenergy/compas/scl/data/repository/CompasSclDataRepository.java
@@ -9,12 +9,8 @@ import org.lfenergy.compas.scl.data.model.SclMetaInfo;
 import org.lfenergy.compas.scl.data.model.Version;
 import org.lfenergy.compas.scl.extensions.model.SclFileType;
 
-import javax.transaction.Transactional;
 import java.util.List;
 import java.util.UUID;
-
-import static javax.transaction.Transactional.TxType.REQUIRED;
-import static javax.transaction.Transactional.TxType.SUPPORTS;
 
 /**
  * Repository class that will be used to handle SCL File for a specific type of storage.
@@ -27,7 +23,6 @@ public interface CompasSclDataRepository {
      * @param type The type of SCL to search for.
      * @return The list of entries found for the passed type.
      */
-    @Transactional(SUPPORTS)
     List<Item> list(SclFileType type);
 
     /**
@@ -37,7 +32,6 @@ public interface CompasSclDataRepository {
      * @param id   The ID of the SCL to search for.
      * @return The list of versions found for that specific sCl Entry.
      */
-    @Transactional(SUPPORTS)
     List<HistoryItem> listVersionsByUUID(SclFileType type, UUID id);
 
     /**
@@ -47,7 +41,6 @@ public interface CompasSclDataRepository {
      * @param id   The ID of the SCL to search for.
      * @return The SCL XML File Content that is search for.
      */
-    @Transactional(SUPPORTS)
     String findByUUID(SclFileType type, UUID id);
 
     /**
@@ -57,7 +50,6 @@ public interface CompasSclDataRepository {
      * @param id   The ID of the SCL to search for.
      * @return The Meta Info of SCL Entry that is search for.
      */
-    @Transactional(SUPPORTS)
     SclMetaInfo findMetaInfoByUUID(SclFileType type, UUID id);
 
     /**
@@ -68,7 +60,6 @@ public interface CompasSclDataRepository {
      * @param version The version of the ScL to search for.
      * @return The SCL XML File Content that is search for.
      */
-    @Transactional(SUPPORTS)
     String findByUUID(SclFileType type, UUID id, Version version);
 
     /**
@@ -78,7 +69,6 @@ public interface CompasSclDataRepository {
      * @param name The name of the SCL used for checking duplicates.
      * @return True if name is already used by another SCL File of the same File type, otherwise false.
      */
-    @Transactional(SUPPORTS)
     boolean hasDuplicateSclName(SclFileType type, String name);
 
     /**
@@ -94,26 +84,24 @@ public interface CompasSclDataRepository {
      * @param scl     The SCL XML File content to store.
      * @param version The version of the new entry to be created.
      * @param who     The user that created the new entry.
+     * @param labels  The list of Labels extracted from the SCL XML File.
      */
-    @Transactional(REQUIRED)
-    void create(SclFileType type, UUID id, String name, String scl, Version version, String who);
+    void create(SclFileType type, UUID id, String name, String scl, Version version, String who, List<String> labels);
 
     /**
-     * Delete all versions for a specific SCL File using it's ID.
+     * Delete all versions for a specific SCL File using its ID.
      *
      * @param type The type of SCL where to find the SCL File
      * @param id   The ID of the SCL File to delete.
      */
-    @Transactional(REQUIRED)
     void delete(SclFileType type, UUID id);
 
     /**
-     * Delete passed versions for a specific SCL File using it's ID.
+     * Delete passed versions for a specific SCL File using its ID.
      *
      * @param type    The type of SCL where to find the SCL File
      * @param id      The ID of the SCL File to delete.
      * @param version The version of that SCL File to delete.
      */
-    @Transactional(REQUIRED)
     void delete(SclFileType type, UUID id, Version version);
 }
