+++ service/pom.xml
@@ -28,6 +28,11 @@ SPDX-License-Identifier: Apache-2.0
             <artifactId>scl-extension</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>jakarta.enterprise</groupId>
+            <artifactId>jakarta.enterprise.cdi-api</artifactId>
+        </dependency>
+
         <!-- Test Dependencies -->
         <dependency>
             <groupId>org.mockito</groupId>
